wip: moved reusable functions to utils/misc
